pyproject.toml
[project]
name = "memory-dump-debugger"
version = "0.1.0"
description = "AI-powered memory dump analyzer using LangGraph and WinDbg"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "azure-ai-inference>=1.0.0b9",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.5.0",
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
]

[project.scripts]
dump-debugger = "dump_debugger.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dump_debugger"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
